Use UTC; parse RealTime in logs as UTC if marked as such.
For times which aren't marked as UTC (i.e. RealTime log entries without the
Z suffix), we assume the current local time zone. This may be wrong, but
I really don't want to mess with DST etc.

No TZ info is stored in old logs, so we're stuffed anyway for games which
cross DST boundaries. We could infer DST change from the preceding game's
start time and length, but I don't think that it's worth it.