Account for bots; mark empty games at log-scan time.
Bots are lumped together as one player.
Stats are recorded, but no stats are shown for them.

Games are regarded as empty if no non-bot players were present and active.
It is therefore most efficient to mark empty games during log scanning,
avoiding a lot of complicated, slow SQL queries.